% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_mutations.R
\name{condense_mutations}
\alias{condense_mutations}
\title{Given a set of level 3 features, simplify and condense into level 1 and level 2 mutations}
\usage{
condense_mutations(features)
}
\arguments{
\item{features}{a vector of level 3 features}
}
\value{
\code{condense_mutations} returns a vector of condensed features
}
\description{
Given a set of level 3 features, simplify and condense into level 1 and level 2 mutations
}
