% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondenseMutations.R
\name{CondenseMutations}
\alias{CondenseMutations}
\title{Given a set of level 3 features, simplify and condense into level 1 and level 2 mutations}
\usage{
CondenseMutations(features)
}
\arguments{
\item{features}{a vector of level 3 features}
}
\value{
\code{CondenseMutations} returns a vector of condensed features
}
\description{
Given a set of level 3 features, simplify and condense into level 1 and level 2 mutations
}
