% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextMatters.R
\name{ContextMatters}
\alias{ContextMatters}
\title{Function for binomial testing using a hierarchical tree structure}
\usage{
ContextMatters(muts_df, p_thresh = 0.05, tot_pseudo = 1000)
}
\arguments{
\item{p_thresh}{alpha level for binomial test (default = 0.05)}

\item{tot_pseudo}{pseudo-count value (default = 1000)}

\item{muts_dif}{dataset of mutations}
}
\value{
output vector of survival mutations that pass
the binomial tree testing
}
\description{
Obtain a list of survival mutations by performing a series of nested binomial 
tests for selecting and pruning significant mutations
}
