% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_level3.R
\name{convert_to_level3}
\alias{convert_to_level3}
\title{Given a set of features, convert them to a union set of fundamental (level 3) 
mutation features}
\usage{
convert_to_level3(features)
}
\arguments{
\item{features}{vector of features}
}
\value{
output vector of features
}
\description{
Given a set of features, convert them to a union set of fundamental (level 3) 
mutation features
}
