% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertToLevel3.R
\name{ConvertToLevel3}
\alias{ConvertToLevel3}
\title{Given a set of features, convert them to a union set of fundamental (level 3) 
mutation features}
\usage{
ConvertToLevel3(features)
}
\arguments{
\item{features}{a vector of features}
}
\value{
\code{ConvertToLevel3} returns a vector of fundamental (level 3) features
}
\description{
Given a set of features, convert them to a union set of fundamental (level 3) 
mutation features
}
