% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_dt.R
\docType{data}
\name{example_dt}
\alias{example_dt}
\title{Example dataset of mutations}
\format{
A data frame with 10 rows and 5 columns:
\describe{
  \item{sample_id}{ID of the patient}
  \item{age}{age of the patient}
  \item{chromosome}{chromosomal position of the mutation}
  \item{position}{position of the mutation}
  \item{ref}{original nucleotide}
  \item{alt}{mutated nucleotide}
}
}
\usage{
example_dt
}
\description{
A dataset containing a list of mutations and other necessary attributes
}
\keyword{datasets}
