% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{FeatureSelection}
\alias{FeatureSelection}
\title{Function for feature selection}
\usage{
FeatureSelection(
  dt,
  test_ind = NULL,
  middle_dt,
  factor,
  keep_nonpredictive = F
)
}
\arguments{
\item{dt}{data frame of mutations}

\item{test_ind}{indices for the test data}

\item{middle_dt}{the middle aged samples (NULL if factor is not age)}

\item{factor}{factor/exposure (e.g. age, smoking)}

\item{keep_nonpredictive}{boolean toggle to combine 
non-predictive features as one feature (default is FALSE)}
}
\value{
output a list of several elements:
features_context_0 is a vector of survival mutations for the unexposed group
features_context_1 is a vector of survival mutations for the exposed group
features_selected is a vector of candidate features ranked by AUC
select_n is the number of top features to retain for each method
dt_new is the transformed data from TransformData
}
\description{
Perform feature selection given dataset of mutations by calling 
a series of other functions to find significant and predictive features
}
