% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMinSigmaAlgebra.R
\name{GenerateMinSigmaAlgebra}
\alias{GenerateMinSigmaAlgebra}
\title{Create the smallest non-overlapping partition that generates 
the minimal sigma algebra containing two sets of mutation features}
\usage{
GenerateMinSigmaAlgebra(input_ls, condense = F)
}
\arguments{
\item{input_ls}{a named list of length 2, where each element is a vector of mutation features}

\item{condense}{boolean toggle for whether the output new_partition should return
mutations in condensed form (default = FALSE)}
}
\value{
output is a list of 1 element, new_partition, which is the partition of features
}
\description{
Create the smallest non-overlapping partition that generates 
the minimal sigma algebra containing two sets of mutation features
}
