% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature.R
\name{get_signature}
\alias{get_signature}
\title{Function to obtain a SuperSig}
\usage{
get_signature(dt, factor, wgs = F)
}
\arguments{
\item{dt}{a data frame of mutations}

\item{factor}{the factor/exposure (e.g. "age", "smoking")}

\item{wgs}{logical value indicating whether sequencing data is 
whole-genome (wgs = \code{TRUE}) or whole-exome (wgs = \code{FALSE})}
}
\value{
\code{get_signature} returns an object of class "SuperSig"
}
\description{
Generate a tissue-specific SuperSig for a given dataset of mutations and
exposure factor. Returns the SuperSig and a classification model trained with
the SuperSig.
}
\examples{
# mutation_data is a data frame of mutation counts (see vignette for details)
get_signature(dt = mutation_data, factor = "age")

}
