% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyAuc.R
\name{MyAuc}
\alias{MyAuc}
\title{Calculate AUC for a column}
\usage{
MyAuc(y, x)
}
\arguments{
\item{y}{an indicator variable column, i.e. the "true value" column}

\item{x}{a column of numeric values, e.g. the projected counts of a feature}
}
\value{
`MyAuc` returns a numeric AUC value
}
\description{
Calculate AUC for one column using the Mann-Whitney U statistic
}
