% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyLDAEnvClassifier.R
\name{MyLDAEnvClassifier}
\alias{MyLDAEnvClassifier}
\title{Classification and correlation using signatures}
\usage{
MyLDAEnvClassifier(
  dt,
  test_ind = NULL,
  factor,
  classifier,
  keep_classifier = F,
  adjusted_formula = F,
  features_selected,
  select_n
)
}
\arguments{
\item{dt}{transformed data from FeatureSelection}

\item{test_ind}{indices for test data}

\item{factor}{factor/exposure (e.g. age, smoking)}

\item{classifier}{classifier method to use for prediction (options are "LDA", "Logit", and/or "RF",
must pass as a vector)}

\item{keep_classifier}{boolean toggle for saving the classifier model (default is FALSE)}

\item{adjusted_formula}{boolean toggle for using the adjusted formula (default is FALSE)}

\item{features_selected}{vector of candidate features ranked by AUC}

\item{select_n}{number of top features to retain for each method}
}
\value{
output a list of several elements:
auc is a vector of AUCs for each classifier
signature is the signature (mean differences or rates), created only when
classifier includes "Logit"
classifier is a list of saved models for each input classifier
}
\description{
Classification and correlation using signatures with the option of three 
different classifiers
}
