% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_signature.R
\name{predict_signature}
\alias{predict_signature}
\title{Function to predict using SuperSig object}
\usage{
predict_signature(object, newdata, factor)
}
\arguments{
\item{object}{an object of class "SuperSig"}

\item{newdata}{a data frame of mutations}

\item{factor}{the factor/exposure (e.g. "age", "smoking")}
}
\value{
\code{predict_signature} returns the original data frame with 
additional columns for the feature counts and classification score
}
\description{
Using a generated SuperSig, predict on a new dataset and return 
predicted probabilities for each observation.
}
\examples{

head(example_dt) # use example data from package
input_dt <- make_matrix(example_dt) # convert to correct format
out <- get_signature(dt = input_dt, factor = "age") # get SuperSig

newdata <- predict_signature(out, newdata = input_dt, factor = "Smoking")
newdata \%>\% select(IndVar, F1, score)

}
