% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_vcf.R
\name{process_vcf}
\alias{process_vcf}
\title{Function to transform VCF object into "matrix" format}
\usage{
process_vcf(vcf)
}
\arguments{
\item{vcf}{a VCF object (from `VariantAnnotation` package)}
}
\value{
\code{process_vcf} returns a data frame of mutations, 
one row per mutation
}
\description{
Transform a VCF object into 
a data frame of trinucleotide mutations with flanking bases
in a wide matrix format. The function assumes that the VCF object
contains only one sample and that each row in rowRanges 
represents an observed mutation in the sample.
}
\examples{

}
