% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperSig-class.R
\docType{class}
\name{SuperSig-class}
\alias{SuperSig-class}
\alias{SuperSig}
\title{An S4 class for SuperSig}
\description{
An S4 class for SuperSig
}
\section{Slots}{

\describe{
\item{\code{Signature}}{data frame of features and their 
difference in mean rates between exposed and unexposed
(or the average rate if the factor is "age")}

\item{\code{Features}}{list of features that comprise the signature and 
their representation in terms of the fundamental (trinucleotide) mutations}

\item{\code{AUC}}{length-one numeric vector of the apparent AUC 
(i.e. not cross-validated)}

\item{\code{Model}}{list of a glm class for the trained 
logistic regression model}
}}

