% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supersig_ls.R
\docType{data}
\name{supersig_ls}
\alias{supersig_ls}
\title{Trained SuperSigs from TCGA}
\format{
A named list with 67 elements, each of which is a `SuperSig`
}
\usage{
supersig_ls
}
\description{
A list containing 67 SuperSigs
}
\keyword{datasets}
