% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{TransformData}
\alias{TransformData}
\title{Transform data frame of mutations}
\usage{
TransformData(dt, new_partition)
}
\arguments{
\item{dt}{a data frame of mutations}

\item{new_partition}{a partition of features from \code{GenerateMinSigmaAlgebra}}
}
\value{
\code{TransformData} returns a transformed data frame of mutations, with
columns corresponding to the candidate features with projected counts
and other necessary columns (\code{IndVar}, \code{AGE}, \code{TOTAL_MUTATIONS},
and \code{DIVISON})
}
\description{
Transform data frame of mutations by projecting counts for candidate features
}
