% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMyAuc.R
\name{AllMyAuc}
\alias{AllMyAuc}
\title{Calculate AUC for every column}
\usage{
AllMyAuc(z, IndVar)
}
\arguments{
\item{z}{Data matrix}

\item{IndVar}{Indicator variable for exposure}
}
\value{
output A list of AUCs, one for every column (feature)
}
\description{
AllMyAuc calls MyAuc to calculate the AUC for every column other than the indicator variable
}
