% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContextMatters.R
\name{ContextMatters}
\alias{ContextMatters}
\title{Function for binomial testing using a hierarchical tree structure}
\usage{
ContextMatters(muts_df, p_thresh = 0.05, tot_pseudo = 1000)
}
\arguments{
\item{muts_df}{a data frame of mutations}

\item{p_thresh}{an optional numeric value for the alpha level used in the binomial test (default is `0.05`)}

\item{tot_pseudo}{an optional numeric value for the pseudo-count value (default is `1000`)}
}
\value{
`ContextMatters` returns a vector of survival mutations that pass
the binomial tree testing
}
\description{
Obtain a list of survival mutations by performing a series of nested binomial 
tests for selecting and pruning significant mutations
}
