% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureSelection.R
\name{FeatureSelection}
\alias{FeatureSelection}
\title{Function for feature selection}
\usage{
FeatureSelection(
  dt,
  factor,
  test_ind = NULL,
  middle_dt = NULL,
  keep_nonpredictive = F
)
}
\arguments{
\item{dt}{a data frame of mutations}

\item{factor}{the factor/exposure (e.g. "age", "smoking")}

\item{test_ind}{an optional vector of indices for the test data}

\item{middle_dt}{a data frame of the middle aged samples (\code{NULL}
if the factor is not age)}

\item{keep_nonpredictive}{logical value indicating whether to combine 
non-predictive features as one feature (default is \code{FALSE})}
}
\value{
\code{FeatureSelection} returns a list of several elements:
\itemize{
\item \code{features_context_0} is a vector of survival mutations
for the unexposed group
\item \code{features_context_1} is a vector of survival mutations
for the exposed group
\item \code{features_selected} is a vector of candidate features ranked by AUC
\item \code{new_partition} is a list of the partitioned candidate features,
where each feature is represented by a vector of fundamental mutations
\item \code{select_n} is the number of top features to retain for each method
\item \code{dt_new} is the transformed data from \code{TransformData}
}
}
\description{
Perform feature selection given a dataset of mutations by calling 
a series of other functions to find significant and predictive features
}
