% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMinSigmaAlgebra.R
\name{GenerateMinSigmaAlgebra}
\alias{GenerateMinSigmaAlgebra}
\title{Generate partition of features}
\usage{
GenerateMinSigmaAlgebra(input_ls, condense = F, partitioned_features = F)
}
\arguments{
\item{input_ls}{a named list of length 2, where each element is a
vector of mutation features}

\item{condense}{an optional logical value indicating whether the output value
`new_partition` should return mutations in condensed form (default is \code{FALSE})}

\item{partitioned_features}{indicates whether the mutation features have been prepartitioned
(i.e. input is coming from output of GenerateMinSigmaAlgebra)}
}
\value{
\code{GenerateMinSigmaAlgebra} returns a list of partitioned features,
where each feature is represented by a vector of the fundamental (level 3)
mutations
}
\description{
Find the smallest non-overlapping partition that generates 
the minimal sigma algebra containing two sets of mutation features
}
