% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature.R
\name{get_signature}
\alias{get_signature}
\title{Function to obtain a SuperSig}
\usage{
get_signature(dt, factor, wgs = F)
}
\arguments{
\item{dt}{a data frame of mutations (see vignette for details)}

\item{factor}{the factor/exposure (e.g. "age", "smoking")}

\item{wgs}{logical value indicating whether sequencing data is 
whole-genome (wgs = \code{TRUE}) or whole-exome (wgs = \code{FALSE})}
}
\value{
\code{get_signature} returns an object of class "SuperSig"
}
\description{
Generate a tissue-specific SuperSig for a given dataset of mutations and
exposure factor. Returns the SuperSig and a classification model trained with
the SuperSig.
}
\examples{

print(example_dt) # use example data 
input_dt = make_matrix(example_dt) # convert to correct format
get_signature(dt = input_dt, factor = "age") # get SuperSig

}
