% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature.R
\name{get_signature}
\alias{get_signature}
\title{Function to obtain a SuperSig}
\usage{
get_signature(dt, factor)
}
\arguments{
\item{dt}{a data frame of mutations}

\item{factor}{the factor/exposure (e.g. "age", "smoking")}
}
\value{
\code{get_signature} returns a list of several elements:
\itemize{
\item \code{Signature} is a vector of the signature, i.e. the features that
comprise the signature and their difference in mean rates (or counts if the factor
is "age")
\item \code{Features} is a list of features that comprise the signature and 
their representation in terms of the fundamental (trinucleotide) mutations
\item \code{AUC} is the apparent AUC (i.e. not cross-validated) obtained within 
the provided data using the generated SuperSig
\item \code{Model} is the logistic regression model trained on the data and
the generated SuperSig
}
}
\description{
Generate a tissue-specific SuperSig for a given dataset of mutations and
exposure factor. Returns the SuperSig and a classification model trained with
the SuperSig.
}
\examples{
get_signature(mutation_data, "age")

}
