% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyAuc.R
\name{MyAuc}
\alias{MyAuc}
\title{Calculate AUC for one column using the Mann-Whitney U statistic}
\usage{
MyAuc(y, x)
}
\arguments{
\item{y}{Indicator variable column, i.e. "true value"}

\item{x}{A feature column}
}
\value{
output AUC value
}
\description{
Calculate AUC for one column using the Mann-Whitney U statistic
}
