% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyLDAEnvClassifier.R
\name{MyLDAEnvClassifier}
\alias{MyLDAEnvClassifier}
\title{Classification of exposure using signatures}
\usage{
MyLDAEnvClassifier(
  dt,
  test_ind = NULL,
  factor,
  classifier,
  keep_classifier = F,
  adjusted_formula = F,
  features_selected,
  select_n
)
}
\arguments{
\item{dt}{a transformed data frame from FeatureSelection}

\item{test_ind}{an optional vector of indices for the test data}

\item{factor}{the factor/exposure (e.g. "age", "smoking")}

\item{classifier}{a vector of the classifier method(s) to use for prediction
(options are "LDA", "Logit", and "RF")}

\item{keep_classifier}{an optional logical value indicating whether to save
the classifier model (default is `FALSE`)}

\item{adjusted_formula}{an optional logical value indicating whether to
use the adjusted formula for non-age factors (default is `FALSE`)}

\item{features_selected}{a vector of candidate features ranked by AUC}

\item{select_n}{the number of top features to retain for each method}
}
\value{
`MyLDAEnvClassifier` returns a list of several elements:
\itemize{
\item `auc` is a vector of AUCs for each classifier
\item `signature` is a vector of mean differences or rates (created only when
`classifier` includes "Logit")
\item `classifier` is a list of saved models for each input classifier (created
only when `keep_classifier` is `TRUE`)
}
}
\description{
Calculate the AUC using signatures with the option of three 
different classifiers
}
