% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictiveFeatures.R
\name{PredictiveFeatures}
\alias{PredictiveFeatures}
\title{Select predictive features}
\usage{
PredictiveFeatures(train, new_partition, factor)
}
\arguments{
\item{train}{a transformed data of mutations from \code{TransformData}}

\item{new_partition}{a partition of features from \code{GenerateMinSigmaAlgebra}}

\item{factor}{the factor/exposure (e.g. "age", "smoking")}
}
\value{
`PredictiveFeatures` returns a list of two elements:
\itemize{
\item \code{features_selected} is a vector of candidate features ranked by AUC
\item \code{select_n} is the number of top features to retain for each method,
chosen by cross-validation
}
}
\description{
Rank candidate features by their median bootstrap AUC and find the top n features
for each classifier method, where n is chosen by cross-validation
}
