% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{TransformData}
\alias{TransformData}
\title{Transform data frame of mutations}
\usage{
TransformData(
  dt,
  features_context_0,
  features_context_1,
  new_partition,
  factor
)
}
\arguments{
\item{dt}{a data frame of mutations}

\item{features_context_0}{a vector of candidate features for the unexposed group
returned by \code{ContextMatters}}

\item{features_context_1}{a vector of candidate features for the exposed group
returned by \code{ContextMatters}}

\item{new_partition}{a partition of features from \code{GenerateMinSigmaAlgebra}}

\item{factor}{the factor/exposure (e.g. "age", "smoking")}
}
\value{
\code{TransformData} returns a transformed data frame of mutations, with
columns corresponding to the candidate features with projected counts
and other necessary columns (\code{IndVar}, \code{AGE}, \code{TOTAL_MUTATIONS},
and \code{DIVISON})
}
\description{
Transform data frame of mutations by projecting counts for candidate features
}
