% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{TransformData}
\alias{TransformData}
\title{Transform mutation data to only contain necessary columns}
\usage{
TransformData(
  dt,
  features_context_0,
  features_context_1,
  new_partition,
  factor
)
}
\arguments{
\item{dt}{data frame of mutations}

\item{features_context_0}{candidate features for unexposed group returned by ContextMatters}

\item{features_context_1}{candidate features for exposed group returned by ContextMatters}

\item{new_partition}{partition of features from GenerateMinSigmaAlgebra}
}
\value{
output transformed data of mutations, with columns corresponding to
selected candidate features with projected counts and other necessary data
}
\description{
Transform mutation data to only contain necessary columns
}
